Implemented Dark Theme
- Added CSS files for dark theme
- Refactored icon drawing functions into icons.ts, the icon functions take an additional parameter `dark` now, that determines if the dark theme is activated. Some icons are drawn with different colors to be more visible on the dark background.
