+++ src/ide/css-dark/codemirror.css
@@ -1,0 +1,193 @@
+/* BASICS */
+
+body.dark-theme *.CodeMirror {
+	color: white;
+}
+
+/* PADDING */
+
+
+/* GUTTER */
+
+body.dark-theme *.CodeMirror-gutters {
+	border-right-color: #333;
+	background-color: #242424;
+}
+
+body.dark-theme *.CodeMirror-linenumber {
+	color: #777;
+}
+
+body.dark-theme *.breakpoints {
+	background: #272727;
+}
+
+body.dark-theme *.CodeMirror-guttermarker {
+	color: #fff;
+}
+body.dark-theme *.CodeMirror-guttermarker-subtle {
+	color: #777;
+}
+
+/* CURSOR */
+
+body.dark-theme *.CodeMirror-cursor {
+	border-left-color: #fff;
+}
+/* Shown when moving in bi-directional text */
+body.dark-theme *.CodeMirror div.CodeMirror-secondarycursor {
+	border-left-color: silver;
+}
+
+body.dark-theme *.CodeMirror-ruler {
+	border-left-color: #555;
+}
+
+/* DEFAULT THEME */
+
+body.dark-theme *.cm-s-default .cm-header {
+	color: #4af;
+}
+body.dark-theme *.cm-s-default .cm-quote {
+	color: #4f4;
+}
+body.dark-theme *.cm-negative {
+	color: #d44;
+}
+body.dark-theme *.cm-positive {
+	color: #292;
+}
+
+body.dark-theme *.cm-s-default .cm-keyword {
+	color: #4af;
+}
+body.dark-theme *.cm-s-default .cm-atom {
+	color: #4fe;
+}
+body.dark-theme *.cm-s-default .cm-number {
+	color: #4f4;
+}
+body.dark-theme *.cm-s-default .cm-def {
+	color: #4af;
+}
+body.dark-theme *.cm-s-default .cm-variable {
+	color: #fff;
+}
+body.dark-theme *.cm-s-default .cm-punctuation {
+	color: #fff;
+}
+body.dark-theme *.cm-s-default .cm-property {
+	color: #fff;
+}
+body.dark-theme *.cm-s-default .cm-operator {
+	color: #fff;
+}
+body.dark-theme *.cm-s-default .cm-variable-2 {
+	color: #4df;
+}
+/*.cm-s-default .cm-variable-3,
+.cm-s-default .cm-type {
+	color: #085;
+}*/
+body.dark-theme *.cm-s-default .cm-comment {
+	color: #888;
+}
+body.dark-theme *.cm-s-default .cm-string {
+	color: #f44;
+}
+body.dark-theme *.cm-s-default .cm-string-2 {
+	color: #f50;
+}
+body.dark-theme *.cm-s-default .cm-meta {
+	color: #bbb;
+}
+body.dark-theme *.cm-s-default .cm-qualifier {
+	color: #bbb;
+}
+body.dark-theme *.cm-s-default .cm-builtin {
+	color: #c4f;
+}
+body.dark-theme *.cm-s-default .cm-bracket {
+	color: #4f4;
+}
+body.dark-theme *.cm-s-default .cm-tag {
+	color: #7f4;
+}
+body.dark-theme *.cm-s-default .cm-attribute {
+	color: #44f;
+}
+body.dark-theme *.cm-s-default .cm-hr {
+	color: #999;
+}
+body.dark-theme *.cm-s-default .cm-link {
+	color: #44f;
+}
+
+body.dark-theme *.cm-invalidchar {
+	color: #f44;
+}
+
+/* Default styles for common addons */
+
+body.dark-theme *div.CodeMirror span.CodeMirror-matchingbracket {
+	color: #8f4;
+}
+body.dark-theme *div.CodeMirror span.CodeMirror-nonmatchingbracket {
+	color: #f44;
+}
+body.dark-theme *.CodeMirror-matchingtag {
+	background: rgba(255, 150, 0, 0.3);
+}
+body.dark-theme *.CodeMirror-activeline-background {
+	background: #333337;
+}
+
+/* STOP */
+
+/* The rest of this file contains styles related to the mechanics of
+   the editor. You probably shouldn't touch them. */
+
+body.dark-theme *.CodeMirror {
+	background: #222;
+}
+
+body.dark-theme *.CodeMirror-gutter-wrapper {
+	background: none !important;
+	border: none !important;
+}
+
+body.dark-theme *.CodeMirror-selected {
+	background: #445;
+}
+body.dark-theme *.CodeMirror-focused .CodeMirror-selected {
+	background: #449;
+}
+body.dark-theme *.CodeMirror-line::selection,
+body.dark-theme *.CodeMirror-line > span::selection,
+body.dark-theme *.CodeMirror-line > span > span::selection {
+	background: #449;
+}
+body.dark-theme *.CodeMirror-line::-moz-selection,
+body.dark-theme *.CodeMirror-line > span::-moz-selection,
+body.dark-theme *.CodeMirror-line > span > span::-moz-selection {
+	background: #449;
+}
+
+body.dark-theme *.cm-searching {
+	background-color: #333;
+	background-color: rgba(255, 255, 0, 0.4);
+}
+
+body.dark-theme *.CodeMirror-dialog {
+	background: #333;
+}
+
+body.dark-theme *.CodeMirror-dialog input {
+	color: #fff;
+	background-color: #222;
+	border-color: #888;
+}
+
+body.dark-theme *.CodeMirror-dialog input:focus {
+	border-color: #fff;
+}
