+++ src/ide/css-dark/comment.txt
@@ -1,0 +1,9 @@
+The files in this folder provide the styles for the dark theme.
+Because all the css is concatenated to a single css, all dark theme
+tags should be prefixed with `body.dark-theme [*]`. Note that the absence
+means that the styles at the given style selector are probably not
+converted to dark theme. Only color properties should be used in the
+dark theme when possible.
+
+text area -> #222 
+button area -> #333
