+++ src/ide/css-dark/documentation.css
@@ -1,0 +1,187 @@
+/* general page styles */
+
+body.dark-theme *#doc-main {
+	color: #fff;
+}
+
+body.dark-theme *#searchtext {
+	background: #222;
+	color: #fff;
+	border-color: #888;
+}
+
+body.dark-theme input:focus {
+	background: #222;
+	color: #fff;
+	border-color: #fff;
+}
+
+body.dark-theme *#sidebar {
+	background: #272727;
+}
+
+body.dark-theme *#version {
+	color: #fff;
+	background: #272727;
+}
+
+body.dark-theme *#search {
+	color: #fff;
+	background: #57f;
+}
+
+body.dark-theme *#tree {
+	color: #fff;
+	background: #272727;
+}
+
+body.dark-theme *#content {
+	color: #fff;
+	background: #222;
+	border-left: solid 5px #347;
+	border-right: solid 5px #347;
+}
+
+body.dark-theme *.related {
+	background: #272727;
+	border-color: #57f;
+}
+
+/* styles for the content */
+
+body.dark-theme a:link,
+body.dark-theme a:visited {
+	color: #57f;
+}
+
+body.dark-theme a:active,
+body.dark-theme a:hover {
+	color: #f44;
+}
+
+body.dark-theme hr {
+	background: #151515;
+}
+
+body.dark-theme *.nicetable th {
+	border: 2px solid #fa8;
+	background: #854;
+}
+
+body.dark-theme *.nicetable td {
+	border: 2px solid #abf;
+	background: #568;
+}
+
+body.dark-theme *.nicetable th a {
+	color: #fff;
+}
+
+body.dark-theme *.nicetable td a {
+	color: #fff;
+}
+
+body.dark-theme *.methods th {
+	border-color: #888;
+	background: #543;
+}
+
+body.dark-theme *.methods td {
+	border-color: #888;
+}
+
+body.dark-theme *.searchresults {
+	color: #666;
+}
+
+/* styles for the tree view */
+
+body.dark-theme *.current {
+	color: #fa4;
+	background: #654;
+}
+
+body.dark-theme *.tgui-tree {
+	color: #fff;
+}
+
+body.dark-theme *.tgui-tree-toggle {
+	color: #fff;
+}
+
+body.dark-theme *.tgui-tree-toggle:hover {
+	color: #fff;
+}
+
+body.dark-theme *.tgui-tree-element:hover {
+	background: #444;
+}
+
+/* highlighting */
+
+body.dark-theme *.example {
+	border-color: #c66;
+}
+
+body.dark-theme *.warning {
+	border-color: #c66;
+	background: #fcc;
+}
+
+body.dark-theme *.keyword {
+	background: #333;
+	border-color: #333;
+}
+
+body.dark-theme *.code {
+	background: #151515;
+}
+body.dark-theme *.code-keyword {
+	color: #4af;
+}
+body.dark-theme *.code-identifier {
+	color: #fff;
+}
+body.dark-theme *.code-integer {
+	color: #4fe;
+}
+body.dark-theme *.code-real {
+	color: #4f4;
+}
+body.dark-theme *.code-string {
+	color: #f44;
+}
+body.dark-theme *.code-operator {
+	color: #fff;
+}
+body.dark-theme *.code-grouping {
+	color: #4f4;
+}
+body.dark-theme *.code-delimiter {
+	color: #fff;
+}
+body.dark-theme *.code-comment {
+	color: #888;
+}
+
+body.dark-theme *.ebnf {
+	background: #151919;
+}
+body.dark-theme *.ebnf-identifier {
+	color: #f44;
+}
+body.dark-theme *.ebnf-literal {
+	color: #4f4;
+}
+body.dark-theme *.ebnf-special {
+	color: #c4f;
+}
+body.dark-theme *.ebnf-operator {
+	color: #fff;
+}
+body.dark-theme *.ebnf-grouping {
+	color: #c80;
+}
+body.dark-theme *.ebnf-delimiter {
+	color: #888;
+}
