+++ src/ide/css-dark/ide.css
@@ -1,0 +1,98 @@
+body.dark-theme *button {
+	border-color: #fff;
+}
+
+body.dark-theme *.ide-main {
+	color: #fff;
+	background: #f00;
+}
+
+body.dark-theme *.ide-panel-area {
+	background: #111;
+}
+
+body.dark-theme *.ide-toolbar {
+	background: #353535;
+	color: #fff;
+	border-bottom: #555;
+	box-shadow: inset 0px -1px 0px #444;
+}
+
+body.dark-theme *.ide-sourcecode {
+	/* overridden in file codemirror.css, class .CodeMirror */
+	background: #222;
+	color: #fff;
+}
+
+body.dark-theme *.ide-stackview {
+	background: #222;
+	color: #fff;
+}
+
+body.dark-theme *.ide-programview {
+	background: #222;
+	color: #fff;
+}
+
+body.dark-theme *.ide-messages {
+	background: #222;
+	color: #fff;
+}
+
+body.dark-theme *.ide-message:hover {
+	background: #449;
+}
+
+body.dark-theme *.ide-errormessage {
+	color: #f44;
+}
+
+body.dark-theme *.ide-program-ancestor {
+	background: #863;
+}
+
+body.dark-theme *.ide-program-current {
+	background: #653;
+}
+
+body.dark-theme *.ide-turtle {
+	background: #333;
+}
+
+.ide-tutorial {
+	overflow-y: auto;
+	background: #fff4e8;
+	width: 100%;
+	height: 100%;
+	padding: 15px;
+	font-family: sans-serif;
+	font-size: 11pt;
+	line-height: 140%;
+	color: #000;
+}
+
+body.dark-theme *.ide-index {
+	color: #44f;
+}
+
+body.dark-theme *.ide-integer {
+	color: #4fe;
+}
+body.dark-theme *.ide-real {
+	color: #4f4;
+}
+body.dark-theme *.ide-string {
+	color: #f44;
+}
+body.dark-theme *.ide-collection {
+	color: #fa4;
+}
+body.dark-theme *.ide-keyword {
+	color: #4af;
+}
+body.dark-theme *.ide-builtin {
+	color: #c4f;
+}
+body.dark-theme *.ide-userclass {
+	color: #fff;
+}
