+++ src/ide/css/tgui.css
@@ -139,6 +139,7 @@
 }
 
 .tgui-list-box {
+	background-color: #fff;
 	box-shadow: 0px 0px 2px #00000030;
 	border-radius: 2px;
 	border-style: solid;
@@ -148,6 +149,7 @@
 }
 
 .tgui-text-box {
+	background-color: #fff;
 	box-shadow: 0px 0px 2px #00000030;
 	border-radius: 2px;
 	border-style: solid;
