+++ src/ide/tgui.ts
@@ -1,3 +1,5 @@
+import { icons } from "./icons";
+
 var interact = require("interactjs");
 
 ///////////////////////////////////////////////////////////
@@ -10,6 +12,7 @@ export let tgui = (function () {
 	// global mapping of hotkeys to handlers
 	let hotkeys = {};
 	let hotkeyElement = null;
+	module.dark_theme = false;
 
 	// normalize the hotkey to lowercase
 	module.normalizeHotkey = function (hotkey) {
@@ -202,6 +205,32 @@ export let tgui = (function () {
 		};
 	};
 
+	function setupCanvasZoom(canvas)
+	{
+		// zoom
+		// TODO: programmaticly detect whenever zoom changes
+		let zoom = 2; // Good enough
+		//let zoom = window.devicePixelRatio;
+		if (zoom > 1) {
+			canvas.width *= zoom;
+			canvas.height *= zoom;
+			let ctx = canvas.getContext("2d");
+			ctx.scale(zoom, zoom);
+		}
+	}
+
+	function renderCanvas(canvas, draw)
+	{
+		if(typeof draw === "string")
+		{
+			console.assert(draw.substring(0, 5) == "icon:");
+			let id = draw.substring(5);
+			icons[id](canvas, module.dark_theme);
+		}
+		else
+			draw(canvas);
+	}
+
 	// Create a canvas icon element with automaticly zoomed contents
 	// if the website is zoomed to 200% then the actual width of the
 	// canvas is twice as large. The draw function does not need to
@@ -223,25 +252,19 @@ export let tgui = (function () {
 		let canvas = module.createElement({
 			type: "canvas",
 			parent: description.parent,
-			classname: description.hasOwnProperty("classname")
+			classname: "tgui-dynamic-canvas " + (description.hasOwnProperty("classname")
 				? description["classname"]
-				: "tgui",
+				: "tgui"),
 			style: style,
 		});
 		canvas.width = description.width;
 		canvas.height = description.height;
 
-		// zoom
-		// TODO: programmaticly detect whenever zoom changes
-		let zoom = 2; // Good enough
-		//let zoom = window.devicePixelRatio;
-		if (zoom > 1) {
-			canvas.width *= zoom;
-			canvas.height *= zoom;
-			let ctx = canvas.getContext("2d");
-			ctx.scale(zoom, zoom);
-		}
-		description.draw(canvas);
+		if(typeof description.draw === "string")
+			canvas.setAttribute("tgui_draw", description.draw);
+
+		setupCanvasZoom(canvas);
+		renderCanvas(canvas, description.draw);
 
 		return canvas;
 	};
@@ -841,21 +864,11 @@ export let tgui = (function () {
 		free_panel_id++;
 
 		if (!control.hasOwnProperty("icondraw"))
-			control.icondraw = function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 1;
-				ctx.fillStyle = "#fff";
-				ctx.fillRect(2.5, 2.5, 15, 15);
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 15, 15);
-				ctx.fillStyle = "#00c";
-				ctx.fillRect(2.5, 2.5, 15, 3);
-				ctx.stroke();
-			};
+			control.icondraw = "icon:window";
 
 		// register the panel
 		module.panels.push(control);
+		
 
 		// create the title bar with buttons
 		control.titlebar_container = tgui.createElement({
@@ -898,22 +911,7 @@ export let tgui = (function () {
 			},
 			width: 20,
 			height: 20,
-			draw: function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 1;
-				ctx.strokeStyle = "#666";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 13, 13);
-				ctx.stroke();
-				ctx.fillStyle = "#fff";
-				ctx.fillRect(2.5, 2.5, 7, 13);
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 7, 13);
-				ctx.fillStyle = "#00c";
-				ctx.fillRect(2.5, 2.5, 7, 3);
-				ctx.stroke();
-			},
+			draw: "icon:dockLeft",
 			parent: control.titlebar_container,
 			classname: "tgui-panel-dockbutton",
 			"tooltip-right": "Dock left",
@@ -925,22 +923,7 @@ export let tgui = (function () {
 			},
 			width: 20,
 			height: 20,
-			draw: function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 1;
-				ctx.strokeStyle = "#666";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 13, 13);
-				ctx.stroke();
-				ctx.fillStyle = "#fff";
-				ctx.fillRect(8.5, 2.5, 7, 13);
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.rect(8.5, 2.5, 7, 13);
-				ctx.fillStyle = "#00c";
-				ctx.fillRect(8.5, 2.5, 7, 3);
-				ctx.stroke();
-			},
+			draw: "icon:dockRight",
 			parent: control.titlebar_container,
 			classname: "tgui-panel-dockbutton",
 			"tooltip-right": "Dock right",
@@ -952,18 +935,7 @@ export let tgui = (function () {
 			},
 			width: 20,
 			height: 20,
-			draw: function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 1;
-				ctx.fillStyle = "#fff";
-				ctx.fillRect(2.5, 2.5, 13, 13);
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 13, 13);
-				ctx.fillStyle = "#00c";
-				ctx.fillRect(2.5, 2.5, 13, 3);
-				ctx.stroke();
-			},
+			draw: "icon:maximize",
 			parent: control.titlebar_container,
 			classname: "tgui-panel-dockbutton",
 			"tooltip-right": "Maximize",
@@ -975,22 +947,7 @@ export let tgui = (function () {
 			},
 			width: 20,
 			height: 20,
-			draw: function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 1;
-				ctx.strokeStyle = "#666";
-				ctx.beginPath();
-				ctx.rect(2.5, 2.5, 13, 13);
-				ctx.stroke();
-				ctx.fillStyle = "#fff";
-				ctx.fillRect(4.5, 5.5, 9, 8);
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.rect(4.5, 5.5, 9, 8);
-				ctx.fillStyle = "#00c";
-				ctx.fillRect(4.5, 5.5, 9, 3);
-				ctx.stroke();
-			},
+			draw: "icon:float",
 			parent: control.titlebar_container,
 			classname: "tgui-panel-dockbutton",
 			"tooltip-right": "Floating",
@@ -1002,15 +959,7 @@ export let tgui = (function () {
 			},
 			width: 20,
 			height: 20,
-			draw: function (canvas) {
-				let ctx = canvas.getContext("2d");
-				ctx.lineWidth = 2.5;
-				ctx.strokeStyle = "#000";
-				ctx.beginPath();
-				ctx.moveTo(3, 14.5);
-				ctx.lineTo(15, 14.5);
-				ctx.stroke();
-			},
+			draw: "icon:minimize",
 			parent: control.titlebar_container,
 			classname: "tgui-panel-dockbutton",
 			"tooltip-right": "Minimize",
@@ -1519,20 +1468,7 @@ export let tgui = (function () {
 					},
 					width: 20,
 					height: 20,
-					draw: function (canvas) {
-						let ctx = canvas.getContext("2d");
-						ctx.lineWidth = 2;
-						ctx.strokeStyle = "#000";
-						ctx.beginPath();
-						ctx.arc(9, 6, 4, 1 * Math.PI, 2.25 * Math.PI, false);
-						ctx.lineTo(9, 11.5);
-						ctx.lineTo(9, 13);
-						ctx.stroke();
-						ctx.beginPath();
-						ctx.moveTo(9, 15);
-						ctx.lineTo(9, 17);
-						ctx.stroke();
-					},
+					draw: "icon:help",
 					classname: "tgui-panel-dockbutton",
 					"tooltip-right": "Help",
 				});
@@ -1545,19 +1481,7 @@ export let tgui = (function () {
 				},
 				width: 20,
 				height: 20,
-				draw: function (canvas) {
-					let ctx = canvas.getContext("2d");
-					ctx.lineWidth = 2;
-					ctx.strokeStyle = "#000";
-					ctx.beginPath();
-					ctx.moveTo(4, 4);
-					ctx.lineTo(14, 14);
-					ctx.stroke();
-					ctx.beginPath();
-					ctx.moveTo(4, 14);
-					ctx.lineTo(14, 4);
-					ctx.stroke();
-				},
+				draw: "icon:close",
 				classname: "tgui-panel-dockbutton",
 				"tooltip-right": "Close",
 			});
@@ -1620,57 +1544,9 @@ export let tgui = (function () {
 		tgui.startModal(dlg);
 	};
 
-	module.msgBoxQuestion = function (canvas) {
-		let ctx = canvas.getContext("2d");
-		ctx.lineWidth = 2;
-		ctx.strokeStyle = "#04d";
-		ctx.fillStyle = "#16f";
-		ctx.beginPath();
-		ctx.arc(20, 20, 18.5, 0, 2 * Math.PI);
-		ctx.closePath();
-		ctx.fill();
-		ctx.stroke();
+	module.msgBoxQuestion = "icon:msgBoxQuestion";
 
-		ctx.strokeStyle = "#fff";
-		ctx.beginPath();
-		ctx.arc(20, 15, 7, 1 * Math.PI, 2.25 * Math.PI, false);
-		ctx.lineTo(20, 25);
-
-		ctx.lineTo(20, 28);
-		ctx.stroke();
-		ctx.beginPath();
-		ctx.moveTo(20, 31);
-		ctx.lineTo(20, 34);
-		ctx.stroke();
-	};
-
-	module.msgBoxExclamation = function (canvas) {
-		let ctx = canvas.getContext("2d");
-		ctx.lineWidth = 2;
-		ctx.strokeStyle = "#a91";
-		ctx.fillStyle = "#ec2";
-		ctx.beginPath();
-		//ctx.arc(20, 20, 18.5, 0, 2*Math.PI);
-		ctx.lineTo(19, 2);
-		ctx.lineTo(21, 2);
-		ctx.lineTo(38, 36);
-		ctx.lineTo(37, 38);
-		ctx.lineTo(3, 38);
-		ctx.lineTo(2, 36);
-		ctx.closePath();
-		ctx.fill();
-		ctx.stroke();
-
-		ctx.strokeStyle = "#000";
-		ctx.beginPath();
-		ctx.moveTo(20, 10);
-		ctx.lineTo(20, 28);
-		ctx.stroke();
-		ctx.beginPath();
-		ctx.moveTo(20, 31);
-		ctx.lineTo(20, 34);
-		ctx.stroke();
-	};
+	module.msgBoxExclamation = "icon:msgBoxExclamation";
 
 	// Show a (newly created) modal dialog, that was created by createModal.
 	// Modal dialogs can be stacked. The dialog should not have been shown yet.
@@ -1810,6 +1686,23 @@ export let tgui = (function () {
 		}
 		return true;
 	});
+
+
+	module.setDarkTheme = function(dark_theme: boolean)	{
+		if(module.dark_theme != dark_theme)
+		{
+			document.body.classList.toggle("dark-theme", dark_theme);
+			module.dark_theme = dark_theme;
+			document.querySelectorAll(".tgui-dynamic-canvas").forEach((canvas : any) => {
+				if(canvas.hasAttribute("tgui_draw")) {
+					let ctx = canvas.getContext("2d");
+					//ctx.resetTransform();
+					ctx.clearRect(0, 0, canvas.width, canvas.height);
+					renderCanvas(canvas, canvas.getAttribute("tgui_draw"));
+				}
+			});
+		}
+	}
 
 	return module;
 })();
